chore: svg 파일을 사용하기 위한 설정 추가
- svg module type 추가
- gatsby-plugin-react-svg 설치
