+++ gatsby-config.ts
@@ -64,6 +64,14 @@ const config: GatsbyConfig = {
         ignore: [`**/components/**`, `**/*.ts`],
       },
     },
+    {
+      resolve: 'gatsby-plugin-react-svg',
+      options: {
+        rule: {
+          include: /assets/, // See below to configure properly
+        },
+      },
+    },
   ],
 };
