+++ package.json
@@ -26,6 +26,7 @@
     "gatsby-plugin-image": "^2.0.0",
     "gatsby-plugin-mdx": "^4.0.0",
     "gatsby-plugin-page-creator": "^4.25.0",
+    "gatsby-plugin-react-svg": "^3.3.0",
     "gatsby-plugin-sharp": "^4.0.0",
     "gatsby-remark-copy-linked-files": "^5.25.0",
     "gatsby-remark-images": "^6.25.0",
